final-pjt-back/requirements.txt
asgiref==3.5.1
asttokens==2.0.5
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==37.0.2
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==3.2.12
django-allauth==0.50.0
django-bootstrap-v5==1.0.11
django-cors-headers==3.12.0
django-extensions==3.1.5
djangorestframework==3.13.1
executing==0.8.3
idna==3.3
ipython==8.3.0
jedi==0.18.1
joblib==1.1.0
matplotlib-inline==0.1.3
numpy==1.22.4
oauthlib==3.2.0
pandas==1.4.2
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.29
pure-eval==0.2.2
pycparser==2.21
Pygments==2.12.0
PyJWT==2.4.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
scikit-learn==1.1.1
scipy==1.8.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.2.0
threadpoolctl==3.1.0
traitlets==5.2.0
urllib3==1.26.9
wcwidth==0.2.5
